problem DeleteDB solved
